refactor: Change dir name
